Fix: Limit not working on execute()
Problem?
Often time in production we don't want to return all matching records instead returning a paginated response based on values of limit and offset is desirable.

What is being fixed?
Even after setting a limit value it was not working and .all() or .execute() function were returning complete set of records matching the search criterion.

How is it fixed?
To break away from loop executing the query a condition is added wherein if total number of fetched records is equal or greater than limit.
If no limit is provided, default value i.e. 10 records will be returned.